+++ base_pack/nfc_magic/scenes/nfc_magic_scene_uscuid_ul_partial_details.c
@@ -1,4 +1,5 @@
 #include "../nfc_magic_app_i.h"
+#include "nfc_magic_scene_partial_details_common.h"
 
 void nfc_magic_scene_uscuid_ul_partial_details_on_enter(void* context) {
     NfcMagicApp* instance = context;
@@ -7,12 +8,8 @@ void nfc_magic_scene_uscuid_ul_partial_details_on_enter(void* context) {
     widget_add_string_element(widget, 0, 0, AlignLeft, AlignTop, FontPrimary, "Not Written");
 
     FuriString* message = furi_string_alloc();
-    // Bitmap is page-ordered, so emit set bits ascending as-is.
-    for(uint16_t page = 0; page < instance->write_progress_total; page++) {
-        if(instance->write_failed_bitmap[page >> 3] & (1u << (page & 7u))) {
-            furi_string_cat_printf(message, "%u ", page);
-        }
-    }
+    nfc_magic_partial_details_append_indices(
+        message, instance->write_failed_bitmap, instance->write_progress_total, 0);
 
     widget_add_text_scroll_element(widget, 0, 13, 128, 51, furi_string_get_cstr(message));
     furi_string_free(message);